changed lib for bot. refactoring...
